OneDrive - wonderdatalabs.com/IRN001-2/requirements.txt
streamlit==1.28.0
numpy==1.26.4
plotly==5.13.0
pillow==10.0.1
python-dotenv==1.0.0
